feat: logic to move to modify ticket on total pending ticket
